+++ src/datasources/bpftrace/types.ts
@@ -1,7 +1,6 @@
-import { TargetFormat } from '../lib/models/pcp';
 import { DataSourceJsonData } from '@grafana/data';
 import { Script } from './script';
-import { PmapiQuery } from '../lib/models/pmapi';
+import { PmapiQuery, TargetFormat } from '../lib/models/pmapi';
 
 export interface BPFtraceOptions extends DataSourceJsonData {
     hostspec?: string;