+++ src/datasources/vector/types.ts
@@ -1,6 +1,5 @@
-import { TargetFormat } from '../lib/models/pcp';
 import { DataSourceJsonData } from '@grafana/data';
-import { PmapiQuery } from '../lib/models/pmapi';
+import { PmapiQuery, TargetFormat } from '../lib/models/pmapi';
 
 export interface VectorOptions extends DataSourceJsonData {
     hostspec?: string;