+++ src/lib/services/PmSeriesApiService.ts
@@ -24,9 +24,9 @@ import {
 } from '../models/api/series';
 import { timeout } from '../utils/timeout';
 import Config from '../../components/search/config/config';
-import { DefaultRequestOptions } from '../../datasources/lib/models/pcp';
 import { defaults } from 'lodash';
 import { NetworkError } from '../models/errors/network';
+import { DefaultRequestOptions } from 'datasources/lib/models/pmapi';
 
 class PmSeriesApiService {
     constructor(